[REF] refactor test in order to be able to reuse common code
