+++ shopinvader_product_stock/tests/test_stock_move.py
@@ -6,31 +6,25 @@
 from random import randint
 from contextlib import contextmanager
 from odoo import api, models
-from odoo.addons.component.tests.common import SavepointComponentCase
-from odoo.addons.queue_job.tests.common import JobMixin
+from .common import StockCommonCase
 
 
-class TestStockMove(SavepointComponentCase, JobMixin):
+class TestStockMove(StockCommonCase):
     """
     Tests for stock.move
     """
 
     def setUp(self):
         super(TestStockMove, self).setUp()
-        self.env = api.Environment(self.cr, self.uid, {})
-        self.shopinvader_backend = self.env.ref('shopinvader.backend_1')
-        self.shopinvader_backend.product_stock_field_id = self.env.ref(
-            "stock.field_product_product_qty_available")
-        self.shopinvader_backend.bind_all_product()
         self.loc_supplier = self.env.ref('stock.stock_location_suppliers')
-        self.product = self.env.ref('product.product_product_8')
         self.picking_type_in = self.env.ref("stock.picking_type_in")
 
     def _create_move(self):
         return self.env['stock.move'].create({
             'name': 'Forced Move',
             'location_id': self.loc_supplier.id,
-            'location_dest_id': self.picking_type_in.default_location_dest_id.id,
+            'location_dest_id':\
+                self.picking_type_in.default_location_dest_id.id,
             'product_id': self.product.id,
             'product_uom_qty': 2.0,
             'product_uom': self.product.uom_id.id,
